% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGFF.R
\name{load_gene_models}
\alias{load_gene_models}
\title{Load a GFF file as a GenomicFeatures object}
\usage{
load_gene_models(filename, seqinfo_ = NULL)
}
\arguments{
\item{filename}{GFF filename}

\item{seqinfo_}{Seqinfo object}
}
\value{
Rmonad wrapped GenomicFeatures object
}
\description{
Load a GFF file as a GenomicFeatures object
}
